revert itemlist filters
